pyproject.toml
[project]
name = "notion-rag"
version = "0.1.0"
description = "Notion-to-VectorDB RAG service with Gemini File Search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.62.0",
    "notion-client>=2.7.0",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["notion_rag"]

[project.scripts]
notion-rag = "notion_rag.cli:main"
